Small fix for systemjs.config.js
With the original solution, Angular works only on the default route page (e.g.: http://localhost12345), but not on other routes (e.g.: http://localhost:12345/Home/About). I found out, that it was caused by bad file targeting in systemjs.config.js. In the original "./" case, all the different routed pages, are looking for the given files, relative to them (e.g.: http://localhost:12345/Home/About is looking for Angular files in http://localhost:12345/Home/About/libs/@angular), however, because of the gulping, these files are placed in a fix /lib folder. It is the same case with Scripts. The app should look for the /Scripts folder, located in the root. If we target these folders, Angular works great on all the routes.